Extracted Interface and refactored Dagger Metrics Injection Module to reflect the changes.
