+++ server/src/main/java/com/hedera/block/server/metrics/MetricsService.java
@@ -18,73 +18,17 @@ package com.hedera.block.server.metrics;
 
 import com.swirlds.metrics.api.Counter;
 import com.swirlds.metrics.api.LongGauge;
-import com.swirlds.metrics.api.Metrics;
-import edu.umd.cs.findbugs.annotations.NonNull;
 
-/**
- * Use member variables of this class to update metric data for the Hedera Block Node.
- *
- * <p>Metrics are updated by calling the appropriate method on the metric object instance. For
- * example, to increment a counter, call {@link Counter#increment()}.
- */
-public class MetricsService {
-
-    private static final String CATEGORY = "hedera_block_node";
-
-    // Live BlockItem Counter
-    private static final Counter.Config LIVE_BLOCK_ITEM_COUNTER =
-            new Counter.Config(CATEGORY, "live_block_items").withDescription("Live BlockItems");
-
-    // Block Persistence Counter
-    private static final Counter.Config BLOCK_PERSISTENCE_COUNTER =
-            new Counter.Config(CATEGORY, "blocks_persisted").withDescription("Blocks Persisted");
-
-    // Subscriber Gauge
-    private static final LongGauge.Config SUBSCRIBER_GAUGE =
-            new LongGauge.Config(CATEGORY, "subscribers").withDescription("Subscribers");
-
-    // Single Block Retrieved Counter
-    private static final Counter.Config SINGLE_BLOCK_RETRIEVED_COUNTER =
-            new Counter.Config(CATEGORY, "single_blocks_retrieved")
-                    .withDescription("Single Blocks Retrieved");
-
-    private final Counter liveBlockItems;
-
-    private final Counter blocksPersisted;
-
-    private final Counter singleBlocksRetrieved;
-
-    private final LongGauge subscribers;
-
+public interface MetricsService {
     /** Update the counter of live block items transiting via the live stream. */
-    public final Counter liveBlockItems() {
-        return liveBlockItems;
-    }
+    Counter liveBlockItems();
 
     /** Update the counter of blocks persisted to storage. */
-    public final Counter blocksPersisted() {
-        return blocksPersisted;
-    }
+    Counter blocksPersisted();
 
     /** Update the counter of single blocks retrieved from storage. */
-    public final Counter singleBlocksRetrieved() {
-        return singleBlocksRetrieved;
-    }
+    Counter singleBlocksRetrieved();
 
     /** Update the gauge of subscribers currently consuming to the live stream. */
-    public final LongGauge subscribers() {
-        return subscribers;
-    }
-
-    /**
-     * Create singleton instance of metrics service to be used throughout the application.
-     *
-     * @param metrics the metrics instance
-     */
-    public MetricsService(@NonNull final Metrics metrics) {
-        this.liveBlockItems = metrics.getOrCreate(LIVE_BLOCK_ITEM_COUNTER);
-        this.blocksPersisted = metrics.getOrCreate(BLOCK_PERSISTENCE_COUNTER);
-        this.singleBlocksRetrieved = metrics.getOrCreate(SINGLE_BLOCK_RETRIEVED_COUNTER);
-        this.subscribers = metrics.getOrCreate(SUBSCRIBER_GAUGE);
-    }
+    LongGauge subscribers();
 }
