+++ server/src/test/java/com/hedera/block/server/config/BlockNodeContextTest.java
@@ -20,6 +20,7 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 import com.hedera.block.server.metrics.MetricsService;
+import com.hedera.block.server.metrics.MetricsServiceImpl;
 import com.swirlds.config.api.Configuration;
 import org.junit.jupiter.api.Test;
 
@@ -28,7 +29,7 @@ class BlockNodeContextTest {
     @Test
     void BlockNodeContext_initializesWithMetricsAndConfiguration() {
         Configuration configuration = mock(Configuration.class);
-        MetricsService metricsService = mock(MetricsService.class);
+        MetricsService metricsService = mock(MetricsServiceImpl.class);
 
         BlockNodeContext context = new BlockNodeContext(metricsService, configuration);
